combining delete load balancers and ec2 instance
